move rtc types to rtc_bindings.py
