+++ facefusion/rtc.py
@@ -1,99 +1,56 @@
 import ctypes
 import time
-from typing import List, Optional, Tuple
+from functools import lru_cache
+from typing import Dict, List, Optional, Tuple
 
-from facefusion.rtc_helper import create_static_download_set
-from facefusion.types import RtcAudioTrack, RtcPeer, RtcVideoTrack
-
-
-RTC_CONFIGURATION = type('RtcConfiguration', (ctypes.Structure,),
-{
-	'_fields_':
-	[
-		('iceServers', ctypes.POINTER(ctypes.c_char_p)),
-		('iceServersCount', ctypes.c_int),
-		('proxyServer', ctypes.c_char_p),
-		('bindAddress', ctypes.c_char_p),
-		('certificateType', ctypes.c_int),
-		('iceTransportPolicy', ctypes.c_int),
-		('enableIceTcp', ctypes.c_bool),
-		('enableIceUdpMux', ctypes.c_bool),
-		('disableAutoNegotiation', ctypes.c_bool),
-		('forceMediaTransport', ctypes.c_bool),
-		('portRangeBegin', ctypes.c_ushort),
-		('portRangeEnd', ctypes.c_ushort),
-		('mtu', ctypes.c_int),
-		('maxMessageSize', ctypes.c_int)
-	]
-})
-
-RTC_PACKETIZER_INIT = type('RtcPacketizerInit', (ctypes.Structure,),
-{
-	'_fields_':
-	[
-		('ssrc', ctypes.c_uint32),
-		('cname', ctypes.c_char_p),
-		('payloadType', ctypes.c_uint8),
-		('clockRate', ctypes.c_uint32),
-		('sequenceNumber', ctypes.c_uint16),
-		('timestamp', ctypes.c_uint32),
-		('maxFragmentSize', ctypes.c_uint16),
-		('nalSeparator', ctypes.c_int),
-		('obuPacketization', ctypes.c_int),
-		('playoutDelayId', ctypes.c_uint8),
-		('playoutDelayMin', ctypes.c_uint16),
-		('playoutDelayMax', ctypes.c_uint16)
-	]
-})
-
-LOG_CB_TYPE = ctypes.CFUNCTYPE(None, ctypes.c_int, ctypes.c_char_p)
+from facefusion.common_helper import is_linux, is_macos, is_windows
+from facefusion.download import conditional_download_hashes, conditional_download_sources
+from facefusion.filesystem import resolve_relative_path
+from facefusion.rtc_bindings import RTC_CONFIGURATION, RTC_PACKETIZER_INIT, init_ctypes
+from facefusion.types import DownloadSet, RtcAudioTrack, RtcPeer, RtcVideoTrack
 
 
-def init_ctypes(rtc_library : ctypes.CDLL) -> bool:
-	if rtc_library:
-		rtc_library.rtcInitLogger.argtypes = [ ctypes.c_int, LOG_CB_TYPE ]
-		rtc_library.rtcInitLogger.restype = None
-		rtc_library.rtcInitLogger(4, LOG_CB_TYPE(0))
+def resolve_binary_file() -> Optional[str]:
+	if is_linux():
+		return 'linux-x64-openssl-h264-vp8-av1-opus-libdatachannel-0.24.1.so'
+	if is_macos():
+		return 'macos-universal-openssl-h264-vp8-av1-opus-libdatachannel-0.24.1.dylib'
+	if is_windows():
+		return 'windows-x64-openssl-h264-vp8-av1-opus-datachannel-0.24.1.dll'
+	return None
 
-		rtc_library.rtcCreatePeerConnection.argtypes = [ ctypes.POINTER(RTC_CONFIGURATION) ]
-		rtc_library.rtcCreatePeerConnection.restype = ctypes.c_int
 
-		rtc_library.rtcDeletePeerConnection.argtypes = [ ctypes.c_int ]
-		rtc_library.rtcDeletePeerConnection.restype = ctypes.c_int
+@lru_cache
+def create_static_download_set() -> Dict[str, DownloadSet]: # TODO: replace once conda package is in place
+	binary_name = resolve_binary_file()
 
-		rtc_library.rtcSetRemoteDescription.argtypes = [ ctypes.c_int, ctypes.c_char_p, ctypes.c_char_p ]
-		rtc_library.rtcSetRemoteDescription.restype = ctypes.c_int
+	return\
+	{
+		'hashes':
+		{
+			'datachannel':
+			{
+				'url': 'https://huggingface.co/bluefoxcreation/libdatachannel/resolve/main/linux-x64-openssl-h264-vp8-av1-opus-libdatachannel-0.24.1.so.hash', # TODO: use url with dynamic binary_name
+				'path': resolve_relative_path('../.assets/binaries/' + binary_name + '.hash')
+			}
+		},
+		'sources':
+		{
+			'datachannel':
+			{
+				'url': 'https://huggingface.co/bluefoxcreation/libdatachannel/resolve/main/linux-x64-openssl-h264-vp8-av1-opus-libdatachannel-0.24.1.so', # TODO: use url with dynamic binary_name
+				'path': resolve_relative_path('../.assets/binaries/' + binary_name)
+			}
+		}
+	}
 
-		rtc_library.rtcAddTrack.argtypes = [ ctypes.c_int, ctypes.c_char_p ]
-		rtc_library.rtcAddTrack.restype = ctypes.c_int
 
-		rtc_library.rtcSendMessage.argtypes = [ ctypes.c_int, ctypes.c_void_p, ctypes.c_int ]
-		rtc_library.rtcSendMessage.restype = ctypes.c_int
+def pre_check() -> bool:
+	download_set = create_static_download_set()
 
-		rtc_library.rtcSetVP8Packetizer.argtypes = [ ctypes.c_int, ctypes.POINTER(RTC_PACKETIZER_INIT) ]
-		rtc_library.rtcSetVP8Packetizer.restype = ctypes.c_int
-
-		rtc_library.rtcChainRtcpSrReporter.argtypes = [ ctypes.c_int ]
-		rtc_library.rtcChainRtcpSrReporter.restype = ctypes.c_int
-
-		rtc_library.rtcSetTrackRtpTimestamp.argtypes = [ ctypes.c_int, ctypes.c_uint32 ]
-		rtc_library.rtcSetTrackRtpTimestamp.restype = ctypes.c_int
-
-		rtc_library.rtcIsOpen.argtypes = [ ctypes.c_int ]
-		rtc_library.rtcIsOpen.restype = ctypes.c_bool
-
-		rtc_library.rtcChainRtcpNackResponder.argtypes = [ ctypes.c_int, ctypes.c_uint ]
-		rtc_library.rtcChainRtcpNackResponder.restype = ctypes.c_int
-
-		rtc_library.rtcGetLocalDescription.argtypes = [ ctypes.c_int, ctypes.c_char_p, ctypes.c_int ]
-		rtc_library.rtcGetLocalDescription.restype = ctypes.c_int
-
-		rtc_library.rtcSetOpusPacketizer.argtypes = [ ctypes.c_int, ctypes.POINTER(RTC_PACKETIZER_INIT) ]
-		rtc_library.rtcSetOpusPacketizer.restype = ctypes.c_int
-
-		return True
-
-	return False
+	if not conditional_download_hashes(download_set.get('hashes')):
+		return False
+	return conditional_download_sources(download_set.get('sources'))
 
 
 def load_library() -> Optional[ctypes.CDLL]:
@@ -208,3 +165,13 @@ def delete_peers(rtc_library : ctypes.CDLL, peers : List[RtcPeer]) -> None:
 			rtc_library.rtcDeletePeerConnection(peer_connection_id)
 
 	peers.clear()
+
+
+def is_peer_connected(rtc_library : ctypes.CDLL, peers : List[RtcPeer]) -> bool:
+	for rtc_peer in peers:
+		video_track_id = rtc_peer.get('video_track')
+
+		if video_track_id and rtc_library.rtcIsOpen(video_track_id):
+			return True
+
+	return False
